Fix turbo custom checkout build errors
