+++ android/src/main/java/com/razorpay/flutter_customui/RazorpayPaymentActivity.java
@@ -9,11 +9,11 @@ import android.widget.RelativeLayout;
 import com.razorpay.PaymentResultWithDataListener;
 import com.razorpay.Razorpay;
 import com.razorpay.PaymentData;
-import com.razorpay.upi.UpiAccount;
 
 import org.json.JSONObject;
 import android.content.Intent;
 
+import java.lang.reflect.Method;
 import java.util.HashMap;
 
 import android.app.AlertDialog;
@@ -88,11 +88,12 @@ public class RazorpayPaymentActivity extends Activity implements PaymentResultWi
                 razorpay.changeApiKey(key);
                 JSONObject newPayLoad =  payload.getJSONObject("payload");
                 newPayLoad.remove(Constants.KEY_ID);
-                UpiAccount upiAccount = RazorpayDelegate.getUpiAccount(payload.getString("upiAccount")) ;
+                Object upiAccount = RazorpayDelegate.getUpiAccount(payload.getString("upiAccount")) ;
                 HashMap<String, Object> payloadMap = new HashMap<>();
                 payloadMap.put("upiAccount", upiAccount);
                 payloadMap.put("payload", newPayLoad);
-                razorpay.submit(payloadMap, RazorpayPaymentActivity.this);
+                Method method = razorpay.getClass().getMethod("submit", HashMap.class, PaymentResultWithDataListener.class);
+                method.invoke(razorpay, payloadMap, RazorpayPaymentActivity.this);
                 return;
             }
             razorpay.submit(payload, RazorpayPaymentActivity.this);
